+++ packages/edit/client/views/popups.js
@@ -2,7 +2,15 @@
 Template.popupCursor_edit.events({
 	'click .btn-editinsert': function(e,tmpl) {
 
+		var btn$ = $(e.target),
+			icon$ = btn$.find('.icon');
+		btn$.addClass('disabled');
+		icon$.addClass('icon-loader');
+
 		Meteor.call('insertPlace', {loc: this.loc}, function(err, placeId) {
+
+			btn$.removeClass('disabled');
+			icon$.removeClass('icon-loader');
 
 			Meteor.subscribe('placeById', placeId, function() {
 
@@ -11,6 +19,7 @@ Template.popupCursor_edit.events({
 				var place = K.placeById(placeId);
 				
 				K.Map.hideCursor();
+				
 				K.Map.addItem( place );
 			});
 		});