+++ packages/import/collections/places.js
@@ -1,6 +1,6 @@
 
 K.extend({
-	findPlacesImportByUserId: function(userId) {
+	findPlacesImportByUser: function(userId) {
 		return Places.find({
 			'import.name': {$exists:true},
 			'userId': userId
@@ -9,7 +9,20 @@ K.extend({
 					import:1
 				},
 				sort: { createdAt: -1 },
-				limit: 20
+				limit: 30
+			})
+		);
+	},
+	findPlacesByImportName: function(userId, name) {
+		return Places.find({
+			'import.name': name,
+			'userId': userId
+		}, _.deepExtend({}, K.filters.placeItem, {
+				fields: {
+					import:1
+				},
+				sort: { createdAt: -1 },
+				limit: 30
 			})
 		);
 	}