+++ packages/import/i18n/en.js
@@ -4,5 +4,6 @@ i18n.map('en', {
 	label_importfile: "Upload a GeoJSON File",
 	label_importname: "Import Name",
 	label_imported: "Imported places",
+	label_imports: "Imports",
 	error_import_formatNotValid: "Format not Valid",
 });