+++ packages/import/i18n/it.js
@@ -4,5 +4,6 @@ i18n.map('it', {
 	label_importfile: "Carica un file GeoJSON",
 	label_importname: "Nome Importazione",	
 	label_imported: "Luoghi Importati",
+	label_imports: "Importazioni",
 	error_import_formatNotValid: "Formato non Valido"
 });