+++ packages/import/server/pubs.js
@@ -1,10 +1,24 @@
 
-Meteor.publish('importsByUserId', function(userId) {
+Meteor.publish('importsByUser', function(userId) {
 	if(this.userId)
 	{
-		var cur = K.findPlacesImportByUserId(userId);
+		var cur = K.findPlacesImportByUser(userId);
 
-		console.log('Pub: importsByUserId', userId, cur.count());
+		console.log('Pub: importsByUser', userId, cur.count());
+
+		return cur;
+	}
+	else
+		this.ready();
+});
+
+Meteor.publish('importsByName', function(name) {
+
+	if(this.userId)
+	{
+		var cur = K.findPlacesByImportName(this.userId, name);
+
+		console.log('Pub: importsByName', name, cur.count());
 
 		return cur;
 	}